Establish Domain trait and move implementors to src/domain/ tree
