+++ src/domain/nasa_esa_iss/mod.rs
@@ -1,1 +1,6 @@
 pub mod adapter;
+
+/// The NASA/ESA ISS domain.
+pub struct NasaEsaIss;
+
+impl crate::domain::Domain for NasaEsaIss {}
