+++ src/domain/procedure/mod.rs
@@ -1,3 +1,8 @@
 pub mod adapter;
 pub mod types;
 mod typst;
+
+/// The procedure domain: preserves the full procedure hierarchy when rendering.
+pub struct Procedure;
+
+impl crate::domain::Domain for Procedure {}
