+++ src/domain/recipe/mod.rs
@@ -1,3 +1,8 @@
 pub mod adapter;
 pub mod types;
 mod typst;
+
+/// The recipe domain.
+pub struct Recipe;
+
+impl crate::domain::Domain for Recipe {}
